Pull up ticks and groupByTracks to the wrapper
To escape child component to modify something for the parent (the wrapper)
